Extract reminder_pipeline.rs from turn.rs (Phase 4 batch 1, #684)
Issue #684 (parent #680, Phase 4) batch 1 — opener PR. Create the new
sibling module `src/agent/loop_run/reminder_pipeline.rs` and move the
reminder-orchestration data type out of `turn.rs`.

Migrated to `reminder_pipeline.rs` (`pub(super)`):

* `struct ReminderCallContext { session_id, model, kind, frame,
    mode_label, active_precautions_summary, touched_files, user_task,
    workspace_root, active_precautions_at_call_time, anvil_score,
    anvil_score_from_current_turn }`
* `impl ReminderCallContext { inputs() -> ReminderInputs<'_>,
    log_outcome(turn_index, &outcome, include_inputs) }`

The `reminder.rs` sibling SSOT (`ReminderInputs` / `ReminderOutcome` /
`build_log_payload`) is unchanged. Only the orchestration that
assembles the context lives in the new module.

`turn.rs` adjustments:

* Removed the `struct ReminderCallContext` + `impl` block (originally
  at line 2186-2233).
* Top-of-file `use super::reminder::{...}` block trimmed to just
  `{self, ReminderOutcome}`; `ReminderCallContext` is now imported
  from `super::reminder_pipeline::ReminderCallContext`. The
  `ReminderInputs` / `build_log_payload` aliases moved with the
  struct to `reminder_pipeline.rs`.

`loop_run.rs` adjustments:

* Registered `mod reminder_pipeline;` with a Phase-4 / Issue-684 doc
  comment.

`CLAUDE.md` File Map updated:

* Refreshed the `scaffold_pipeline.rs` entry to reflect the
  consolidated state after batches 1-17 (all dispatch fns moved, no
  veneer shims remaining on `impl Agent`).
* Added the new `reminder_pipeline.rs` entry.

LOC delta:

* `turn.rs`: 30,218 → 30,168 (-50 LOC)
* `reminder_pipeline.rs`: 0 → 88 (new)
* Cumulative `turn.rs`: 39,489 → 30,168 (-9,321 LOC, -23.6%)

`pub(super)` limited / no facade re-export (DR3-001).
`turn.rs` remains the only in-crate consumer.

Phase 4 dispatch methods (`maybe_invoke_reminder` /
`prepare_reminder_context` / `try_clone_reminder_client` /
`run_reminder_sidecar` / friends) stay in `turn.rs` for now and will
migrate in follow-up batches.

Verification:

* `cargo fmt`
* `cargo clippy --lib --all-targets -- -D warnings` clean
* `cargo test --lib` 3,114 passed / 0 failed
